requirements.txt
numpy
torch
Pillow
opencv_python
torchvision
pickle
matplotlib
tqdm